Visitor refactoring, supporting edges.
